Use one CaptionedImage and share Caption with VideoEmbed
