+++ common/styles/components/_captioned_image.scss
@@ -53,66 +53,3 @@
 .captioned-image__image-container {
   position: relative;
 }
-
-.captioned-image__caption {
-  @include clearfix;
-  display: inline-flex; // this style is required to fix a firefox bug, no idea why it works: https://github.com/wellcometrust/wellcomecollection.org/pull/906
-  color: color('black');
-  padding-top: #{($v-spacing-unit * 4) - 8px}; // - to make to adjust for line-height, want padding from x-height
-  padding-bottom: #{($v-spacing-unit * 4) - 8px};
-
-  .icon__shape {
-    fill: color('currentColor');
-  }
-
-  /* stylelint-disable plugin/selector-bem-pattern */
-  p {
-    display: inline;
-
-    &:last-child {
-      margin-bottom: 0;
-    }
-  }
-  /* stylelint-enable plugin/selector-bem-pattern */
-}
-
-.captioned-image__number {
-  display: inline-block;
-  border-right: 1px solid color('pumice');
-}
-
-.captioned-image__caption-text {
-  overflow: hidden;
-  padding-right: map-get($gutter-width , 'small');
-
-  @include respond-to('medium') {
-    padding-right: map-get($gutter-width , 'medium');
-  }
-
-  @include respond-to('large') {
-    padding-right: map-get($gutter-width , 'large');
-  }
-}
-
-.captioned-image__caption-text--truncate {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.captioned-image__truncate-control {
-  background: none;
-  color: color('green');
-  border: 0;
-  padding: $v-spacing-unit 0;
-  margin-left: 2.1em;
-  cursor: pointer;
-
-  &:focus {
-    text-decoration: underline;
-  }
-
-  &.captioned-image__truncate-control--is-hidden {
-    display: none;
-  }
-}
