+++ common/styles/components/_image_gallery.scss
@@ -41,13 +41,6 @@ $image-gallery-control-margin: 18px;
     display: table;
   }
 
-  .captioned-image__caption {
-    display: table-caption;
-    caption-side: bottom;
-    padding-bottom: 0;
-    border-bottom: 0;
-  }
-
   .image {
     max-height: 70vh;
   }
