+++ common/styles/utilities/compiled_variables/_fonts.scss
@@ -25,7 +25,7 @@ $fonts: (
     'font-family-subset': "'Wellcome Bold Web Subset'",
     'font-family-web': "'Wellcome Bold Web'",
     'font-size': 33px,
-    'line-height': 48px
+    'line-height': 42px
   ),
   'WB5': (
     'font-family-base': "'Arial Black', Gadget, sans-serif",
