+++ common/views/components/ImageGallery/ImageGallery.js
@@ -25,7 +25,7 @@ const ImageGallery = ({id, title, items}: ImageGalleryProps) => {
               sizesQueries={'(max-width: 600px) 100vw, ' + (captionedImage.image.width / captionedImage.image.height) * 640 + 'px'}
               preCaptionNode={
                 <span
-                  className={`captioned-image__number ${font({s: 'HNM5'})} ${spacing({s: 2}, {padding: ['right'], margin: ['right']})}`}
+                  className={`inline-block border-right-width-1 border-color-pumice ${font({s: 'HNM5'})} ${spacing({s: 1}, {padding: ['right'], margin: ['right']})}`}
                   aria-label={`slide ${i + 1} of ${items.length}`}>
                   <span aria-hidden='true'>{i + 1}/{items.length}</span>
                 </span>
