+++ server/views/pages/search.njk
@@ -118,7 +118,7 @@
             </div>
           </div>
           <div class="{{ {s: 12, m: 8, l: 5, xl: 5} | gridClasses }}">
-            {% componentJsx 'CaptionedImageNew', henryImage %}
+            {% componentJsx 'CaptionedImage', henryImage %}
           </div>
         </div>
       </div>
